Fix browser/Expo bundle: move package model loading off the Configuration
The packages feature made Configuration statically import node:fs/promises and
require-context (which pulls in node-dir → fs/path). Configuration is bundled for
the browser/Expo, so Metro and the browser-bundle spec failed to resolve those.

- Move the filesystem-backed package model loading into the Node environment
  handler (initializePackageModels), with a no-op in the base handler; Configuration
  just calls getEnvironmentHandler().initializePackageModels(this). node.js is not
  browser-bundled, so require-context/InitializerFromRequireContext live there.
- Make VelociousPackage browser-safe: derive the package directory from the module
  url with the global URL API instead of importing node path/url, so Configuration
  can keep importing it for the ctor.

Configuration now has no static Node-only imports. Browser-bundle spec + the
packages spec pass (7 tests); eslint + typecheck + fallow + build clean.
